More specific type parameters for Codim* aliases
Doing this to Codim1Sweep was necessary to avoid ambiguous method error.
The compiler didn't infer that Codim1Sweep was more specific than
BifurcationSweep when tkind type parameter for Codim1Sweep was
unconstrained.  It's not problem for these type aliases but probably
better to fix it now to avoid future bugs.